// src/types/general.type.ts
export type DaysWeekType = 'monday' | 'tuesday' | 'wednesday' | 'thursday' | 'friday' | 'saturday' | 'sunday';

export enum PriorityType {
    LOW = 'low',
    MEDIUM = 'medium',
    IMPORTANT = 'important',
}

export enum FilterLabelsType{
    ALL = 'all',
    IMPORTANT = 'important',
    COMPLETED = 'completed',
    WAITNG = 'waiting',
    ARCHIVED = 'archived',
    UNARCHIVED = 'unarchived',
}